+++ src/hooks/index.ts
@@ -1,7 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import AsyncStorage, {AsyncStorageStatic} from '@react-native-async-storage/async-storage'
 import {delay} from 'bluebird'
-import {mapValues} from 'lodash'
 import * as React from 'react'
 import {
   onlineManager,
@@ -17,7 +16,6 @@ import {
 import {isNightly} from '../legacy/config'
 import {ObjectValues} from '../legacy/flow'
 import {HWDeviceInfo} from '../legacy/ledgerUtils'
-import {processTxHistoryData} from '../legacy/processTransactions'
 import {WalletMeta} from '../legacy/state'
 import {useStorage} from '../Storage'
 import {parseWalletMeta} from '../Storage/migrations/walletMeta'
@@ -36,7 +34,6 @@ import {generateShelleyPlateFromKey} from '../yoroi-wallets/cardano/shelley/plat
 import {
   Quantity,
   TokenInfo,
-  Transaction,
   TRANSACTION_DIRECTION,
   TRANSACTION_STATUS,
   YoroiAmounts,
@@ -475,26 +472,13 @@ export const useSignTxWithHW = (
   }
 }
 
-const getTransactionInfos = (wallet: YoroiWallet) =>
-  mapValues(wallet.transactions, (tx: Transaction) => {
-    if (!wallet.networkId) throw new Error('invalid state')
-    return processTxHistoryData(
-      tx,
-      wallet.rewardAddressHex != null
-        ? [...wallet.internalAddresses, ...wallet.externalAddresses, ...[wallet.rewardAddressHex]]
-        : [...wallet.internalAddresses, ...wallet.externalAddresses],
-      wallet.confirmationCounts[tx.id] || 0,
-      wallet.networkId,
-    )
-  })
-
 export const useTransactionInfos = (wallet: YoroiWallet) => {
-  const [transactionInfos, setTransactionInfos] = React.useState(() => getTransactionInfos(wallet))
+  const [transactionInfos, setTransactionInfos] = React.useState(() => wallet.transactions)
   React.useEffect(() => {
     const unsubscribe = wallet.subscribe((event) => {
       if (event.type !== 'transactions') return
 
-      setTransactionInfos(getTransactionInfos(wallet))
+      setTransactionInfos(wallet.transactions)
     })
     return () => unsubscribe?.()
   }, [wallet])
