+++ src/yoroi-wallets/cardano/shelley/index.ts
@@ -1,3 +1,2 @@
 export * from './delegationUtils'
 export * from './plate'
-export * from './transactionCache'
